GUI board sq select wired in
GUI board sq, tower, and wall select is wired in.

GameState/GameStateInterface modified and documented to include selector
methods, as well as variables for currently selected wall, tower, and sq
indices.  These variables are initialized to -1 and reset to -1 after
each turn and when selecting another board square.

TODO: **IMPORTANT**

Somebody needs to modify the new card methods in GameState to check for
indices of -1 for selected monster, keep, tower, sq indices when
invoked; for example, the playHitCard method checks if
selectedMonsterIndex is -1, and returns without acting if it is.

Currently, methods like playTimeSlap do not check for -1 (no selected
target index); playing these cards without selecting the appropriate
target WILL crash the game.