Effect cards won't play w/o valid selected index
Made it so the playEffectCard() method waits to see if the particular
effect method corresponding to the particular effect card in question
returns an error (-1) before discarding the card. If the particular
effect method returns -1---meaning for instance that a monster wasn't
selected or a tower wasn't selected---then the card isn't played or
discarded.